Update docker-php-ext-enable
If an uncaught exception is thrown during parsing of the php.ini file, `$extDir` will contain a combination of the exception and the path.  This script shouldn't care about error messages though as long as it gets the path.  This updates the script so that php won't emit any error messages.